Added Rob's Changes to fix the version check.
